% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{result-accessors}
\alias{result-accessors}
\alias{stageTable}
\alias{episodeTable}
\alias{candidateTable}
\alias{diagnosticTable}
\alias{decisionParameters}
\title{Extract result tables and parameters}
\usage{
stageTable(x)

episodeTable(x)

candidateTable(x)

diagnosticTable(x)

decisionParameters(x)
}
\arguments{
\item{x}{A \linkS4class{TransientDTUResult} object.}
}
\value{
\code{stageTable()}, \code{episodeTable()}, \code{candidateTable()}, and
\code{diagnosticTable()} return \link[S4Vectors:DataFrame-class]{S4Vectors::DataFrame} objects.
\code{decisionParameters()} returns a named list.
}
\description{
Extract result tables and parameters
}
\examples{
example <- simulateTransientDTU(n_genes = 8, seed = 10)
fit <- runTransientDTU(example$pairwise, stage_order = example$stage_order)
head(episodeTable(fit))
decisionParameters(fit)$q_threshold

}
